requirements.txt
yfinance
Pillow
